+++ src/main/java/hsrmod/monsters/WarpTrotter.java
@@ -24,7 +24,7 @@ public class WarpTrotter extends BaseMonster implements PreBreakSubscriber {
     float effectTimer = 0.0F;
     float effectInterval = 0.5F;
     
-    public WarpTrotter(int x, int y, int turnCount) {
+    public WarpTrotter(float x, float y, int turnCount) {
         super(ID, 0F, -15.0F, 200, 200, x, y);
         this.turnCount = turnCount;
         effectImages = new TextureAtlas.AtlasRegion[12];
@@ -43,7 +43,7 @@ public class WarpTrotter extends BaseMonster implements PreBreakSubscriber {
         effectImages[11] = new TextureAtlas.AtlasRegion(ImageMaster.loadImage(String.format(url, 12)), 0, 0, 82, 55); //
     }
     
-    public WarpTrotter(int x, int y) {
+    public WarpTrotter(float x, float y) {
         this(x, y, 0);
     }
